Fix tare type quantity to double

+++ Microprojects.Edm.Ui.Logistics/Models/TareType.cs
@@ -5,5 +5,5 @@ public class TareType : DirectoryEntry
     public string Units { get; set; } = string.Empty;
     public bool Countable { get; set; } = true;
     public int Dimensions { get; set; } = 1;
-    public int Capacity { get; set; } = 1;
+    public double Capacity { get; set; } = 1;
 }
